Add files via upload
In the pipeline file, we bring in the data from the database, preprocess the data, and export a csv file ready for ML. 

In the 'Beverly_Grove_Unsupervised_Learning' file, we import that exported file, and run a gamut of supervised machine learning methods. 

in 'Beverly_Grove_Neural_Nets' we use the same file as earlier, but we are using a neural net.